# Base64 faster encode