This pr should be rebased after #1155 has landed.
I'll post benchmark comparisons soon.