@fabianfett 2025-02-04
To analyze performance, we look at the fastest runs, as we can assume that CPU cycles that were not related to the base64 job were minimal in these situations. This allows us to make the most accurate comparisons.

- encode-1MB-toData-noOptions
  - Baseline (in Foundation today): 2894μs
  - Swift implementation: 723μs (4x faster)
  - This PR: 494μs (5.8x faster)

- base64-encode-1MB-toString-lineLength64
  - Baseline (in Foundation today): 3150μs
  - Swift implementation: 708μs (4.4x faster)
  - This PR: 497μs (6.3x faster)

- base64-encode-1MB-toString-noOptions
  - Baseline (in Foundation today): 2915μs
  - Swift implementation: 635μs (4.6x faster)
  - This PR: 414μs (7x faster)

- base64-encode-jwtHeader-toString-noOptions
  - Baseline (in Foundation today): 409ns
  - Swift implementation: 149ns (2.7x faster)
  - This PR: 120ns (3.4x faster)

I think that looks good!